+++ test/_workspace/typescript-parser/async.ts
@@ -1,4 +1,5 @@
 class Class1 {
+    public static async staticAsync() { }
     public func(): void { }
     public async asyncFunc(): Promise<void> { }
     public async asyncFuncWithoutType() { }
@@ -7,3 +8,4 @@ class Class1 {
 function func(): void { }
 async function asyncFunc(): Promise<void> { }
 async function asyncFuncWithoutType() { }
+function promiseFunc(): Promise<void> { }
